pyproject.toml
[project]
name = "raft"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "numpy (>=2.3.2,<3.0.0)",
    "scipy (>=1.16.1,<2.0.0)",
    "pydantic (==2.*)",
    "loguru (>=0.7.3,<0.8.0)",
    "chromadb (>=1.0.15,<2.0.0)",
    "psycopg[binary] (>=3.2.9,<4.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "z3-solver (>=4.15.1.0,<5.0.0.0)",
    "gitpython (>=3.1.0,<4.0.0)",
    "fastapi (>=0.104.0,<1.0.0)",
    "uvicorn[standard] (>=0.24.0,<1.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-mock = "^3.14.1"
black = "^25.1.0"
isort = "^6.0.1"
pre-commit = "^4.2.0"
requests = "^2.32.4"

[tool.poetry]
packages = [{ include = "agent" }]